Modified pricing page and added api routes for pro and basic plan
