+++ src/app/api/checkout-pro/route.ts
@@ -1,0 +1,36 @@
+import { NextRequest, NextResponse } from 'next/server';
+import Stripe from 'stripe';
+
+const stripe = new Stripe(process.env.STRIPE_SECRET_KEY!, {
+  apiVersion: '2023-10-16',
+});
+
+export async function POST(request: NextRequest) {
+  try {
+    const origin = request.headers.get('origin') || 'http://localhost:3000';
+    
+    const session = await stripe.checkout.sessions.create({
+      line_items: [
+        {
+          price_data: {
+            currency: 'eur',
+            product_data: {
+              name: 'Pro Plan',
+            },
+            unit_amount: 20000, // €200.00
+          },
+          quantity: 1,
+        },
+      ],
+      mode: 'payment',
+      success_url: `${origin}/?success=true`,
+      cancel_url: `${origin}/?canceled=true`,
+      automatic_tax: {enabled: true},
+    });
+
+    return NextResponse.json({ url: session.url });
+  } catch (err: any) {
+    console.error('Stripe API Error:', err);
+    return NextResponse.json({ error: err.message }, { status: err.statusCode || 500 });
+  }
+}
